[Webview]: Convert jint to jlong for native pointers in GLFunctionTable to support 64 bit.
Also roll AOSP to include https://android-review.googlesource.com/#/c/88746/

BUG=354405

Review URL: https://codereview.chromium.org/210023002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@259034 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b8963e42e485a82726076dbfd100023e2cdcdea8

+++ android_webview/browser/gpu_memory_buffer_factory_impl.cc
@@ -68,7 +68,7 @@ class GpuMemoryBufferImpl : public gfx::GpuMemoryBuffer {
   }
 
  private:
-  int buffer_id_;
+  long buffer_id_;
   gfx::Size size_;
   bool mapped_;
